Feature-Validate Create Account Data-pt159308568
There was a need to ensure that the data being sent to the server is processed properly before saving to the database.
This commit checks to see if both the password and confirm password matches, then hashes the password. Then with the help of Joi, it checks all fields and ensures they pass a validation test before saving to the database.
The sting fields have now been modiied to accept up to 255 characters.
Files modified are:
- package.json
- controllers/UserController.js
- models/User.js
- models/CreateSchema.js

Feature: Pt159308568