Feature-Add login endpoint to database-pt159324776
This commit adds the function of logging a user. When a user sends their email and password to the server, the email is checked in the database.
If it exists its hash is compared to the hash of the sent password. If it matches, then the user will be logged in. Else the user gets a 401 error.
The test cases have also been modified to relect this change.
N.B: The field ‘dob’ in users table has been changed to ‘dateOfBirth’.
The following files were changed:
- models/User.js
- models/CreateSchema.js
- routes/usersApi.js
- controllers/UserController.js
- tests/users-spec.js

Feature: Pt59324776