Feature-Implement-JWT-#159308791
This commit implements the use of Json Web Token to secure all the endpoints in the api. This only allows a user who has been logged in to access them. Trying to access any route without login will give a 401 error.
The following files are added:
/helpers/checkAuth.js
The following files were modified:
/models/User.js
/routes/usersApi.js
/routes/entriesApi.js

Feature: #159308791