Feature-Get all entries in database-#159308676
This comimit adds the functionality to view all entries for a given user that is logged in. This requires that a user first be signed in and authenticated before accessing this route.
N.B: The token being sent to the browser has now been moved to the response header from the body to accomodate get requests.
The following files were changed:
- controllers/EntryController.js
- controllers/UserController.js
- helpers/checkAuth.js
- models/Entry.js

Feature: #159308676