Feature-Update entry in database-#159324921
This commits adds an endpoint through which a user can make changes and update an entry in the diary. This endpoint can only be accessed if the user is authenticated. Also the entry to be updated must belong to the user.

If a user tries to update an entry 24hrs after posting it, the user would be unable to do so. Moment.js has been added for date formatting.
The following files were modified:
- controllers/EntryController.js
- models/Entry.js
- package.json

Feature: #159324921