Feedback-Implement-feedback-#159392975
This commint implements a number of feedback.
There are now seperate databases for tests and for production/development.  The test database is dropped and created upon running.
The reponse messages have been modified to be more descriptive.
The token has been moved to the body of the response.
The test cases have been modified to match the new responses and make use of the token.
The following files were modified:
- controllers/EntryController.js
-controllers/UserController.js
- helpers/chechAuth.js
- helpers/requests.js
- models/Entry.js
- models/User.js
- tests/entries-spec.js
- tests/global-spec.js
- tests/users-spec.js
- package.json
- app.js

Feedback: #159392975